Extract checksum & cache helpers to AutoUpdater+Checksum
Move cache destination management and SHA-256 verification logic from AutoUpdater.swift into a new AutoUpdater+Checksum.swift extension, following the project's single-responsibility principle. Includes cachedZipDestination, clearCachedDefaults, and verifyChecksum function. Update AutoUpdater.swift documentation to reflect the split.
